+++ src/index.d.ts
@@ -16,6 +16,7 @@ export * from "./provider.js";
 export * from "./runtime-placement.js";
 export * from "./provider-data.js";
 export * from "./provider-runtime.js";
+export * from "./provider-docker.js";
 export * from "./data-plane.js";
 export * from "./forms.js";
 export * from "./ui.js";